updated grade reducer
