+++ app/components/grades/constants.js
@@ -1,3 +1,5 @@
-export const DISPLAY_EXAM_TABLE = 'DISPLAY_EXAM_TABLE'
-export const OPEN_CLASS_LIST = 'OPEN_CLASS_LIST'
-export const UPDATE_EXAM_TABLE = 'UPDATE_EXAM_TABLE'
+export const actions = {
+	DISPLAY_EXAM_TABLE: 'DISPLAY_EXAM_TABLE',
+	OPEN_CLASS_LIST: 'OPEN_CLASS_LIST',
+	UPDATE_EXAM_TABLE: 'UPDATE_EXAM_TABLE'
+}
