+++ app/reducers/index.js
@@ -16,7 +16,7 @@ import {
 } from '../components/subject/reducers'
 import applyFilteredExam from '../components/exam/reducer'
 import { applyStudentData } from '../components/students/reducers'
-import applyGradeData from '../components/grades/reducer'
+import { applyGradeData } from '../components/grades/reducers'
 import { applyMenuStyling } from '../components/sidemenu/reducers'
 import { applyGraphData } from '../components/graphs/reducers'
 import { applyNotesData } from '../components/notes/reducers'
